+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ## Next
 
+- feat: added `browser_resize(width, height)` tool for Playwright
 - refactor: `browser_console_messages(clear, offset, limit, reverse)` and `get_devserver_logs(name, offset, limit, reverse)` can now be controlled better to limit number of returned logs and to get what you actually need (first or last logs).
 - docs: document `autostart` and `prefix_logs` at least with one example in README